Refactor y abm de pago
