+++ src/main/kotlin/com/estonianport/agendaza/dto/EventoDto.kt
@@ -2,6 +2,7 @@ package com.estonianport.agendaza.dto
 
 import java.time.LocalDateTime
 
-class EventoDto(var id: Long, var nombre: String, var codigo : String, var inicio : LocalDateTime, var fin : LocalDateTime) {
+class EventoDto(var id: Long, var nombre: String, var codigo : String,
+                var inicio : LocalDateTime, var fin : LocalDateTime, var tipoEvento : String) {
 
 }
