+++ src/main/kotlin/com/estonianport/agendaza/dto/PagoDto.kt
@@ -1,0 +1,7 @@
+package com.estonianport.agendaza.dto
+
+import java.time.LocalDateTime
+
+class PagoDto(val id: Long, val monto : Int, val codigo : String, val medioDePago: String,
+              val nombreEvento: String, val fecha: LocalDateTime) {
+}
