+++ src/main/kotlin/com/estonianport/agendaza/model/MedioDePago.kt
@@ -1,8 +1,3 @@
 package com.estonianport.agendaza.model
 
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-enum class MedioDePago { EFECTIVO, TARJETA, TRANFERENCIA }
+enum class MedioDePago { EFECTIVO, TRANFERENCIA, TARJETA_DE_DEBITO, TARJETA_DE_CREDITO, CHEQUE, OTRO }
